Smoothed drag scrolling

+++ src/main/java/com/uberverse/arkcraft/common/container/scrollable/IGuiScrollable.java
@@ -2,7 +2,7 @@ package com.uberverse.arkcraft.common.container.scrollable;
 
 public interface IGuiScrollable
 {
-	public int getScrollPosition();
+	public float getScrollPosition();
 
 	public boolean canScroll();
 }
